requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
billiard==3.3.0.23
boto3==1.16.47
botocore==1.19.47
celery==3.1.26.post2
certifi==2020.12.5
chardet==4.0.0
Django==3.1.3
django-celery==3.3.1
django-cors-headers==3.5.0
django-resized==0.3.11
django-storages==1.11.1
djangorestframework==3.12.2
idna==2.10
jmespath==0.10.0
kombu==3.0.37
Pillow==8.0.1
psycopg2==2.8.6
Pygments==2.7.2
python-dateutil==2.8.1
python-resize-image==1.1.19
pytz==2020.4
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
Unidecode==1.1.2
urllib3==1.26.2
whitenoise==5.2.0
